cli: Add config for `LogHistory`
The new log indexer is incredibly useful and powerful, to quickly scan
through logs.

The default setting, 2.35M blocks, isn't optimal in some cases. We'd
want to set it to 0 on some nodes to have unlimited indexing.

This adds the configuration to the `CLI` part (as a flag, and in the
config).