Refactor and optimise commit interrupt (#1570)
Use a simple atomic flag for handling commit interrupt instead of using context based timeout. Internal benchmarks revealed that checking for `context.Done()` for every OPCODE leads to 40% CPU usage (out of the total CPU used in block building loop) which is not very optimal. Instead of context, a pointer to this atomic flag is passed and is toggled when the block building time is completed to interrupt. 

This also allows us to simply the use of commit interrupt and the worker - evm interaction.